Refactored level loading and layered the background. Added pipes and clouds
