+++ public/js/Level.js
@@ -1,15 +1,16 @@
 import Compositor from "./Compositor.js";
 import TileCollider from './TileCollider.js';
-import {Matrix} from './math.js';
 
 export default class Level {
     constructor() {
         this.comp = new Compositor();
         this.entities = new Set();
         this.gravity = 1500;
-        this.tiles = new Matrix();
         this.totalTime = 0;
-        this.tileCollider = new TileCollider(this.tiles);
+        this.tileCollider = null;
+    }
+    setCollisionGrid(matrix) {
+        this.tileCollider = new TileCollider(matrix);
     }
     update(deltaTime) {
         this.entities.forEach(entity => {
