+++ public/js/layers.js
@@ -1,3 +1,5 @@
+import TileResolver from "./TileResolver.js";
+
 export function createCameraLayer(cameraToDraw) {
     return (context, fromCamera) => {
         context.strokeStyle = 'purple';
@@ -11,9 +13,8 @@ export function createCameraLayer(cameraToDraw) {
     }
 }
 
-export function createBackgroundLayer(level, sprites) {
-    const tiles = level.tiles;
-    const resolver = level.tileCollider.tiles;
+export function createBackgroundLayer(level, tiles, sprites) {
+    const resolver = new TileResolver(tiles);
 
     const buffer = document.createElement('canvas');
     buffer.width = 256 + 16;
@@ -21,6 +22,8 @@ export function createBackgroundLayer(level, sprites) {
     const context = buffer.getContext('2d');
 
     function redraw(startIndex, endIndex) {
+        context.clearRect(0, 0, buffer.width, buffer.height);
+
         for (let x=startIndex; x<=endIndex; ++x) {
             const col = tiles.grid[x];
             if (col) {
