+++ public/sprites/overworld.json
@@ -1,49 +1,86 @@
 {
-    "imageURL": "img/tiles.png",
-    "tileW": 16,
-    "tileH": 16,
-    "tiles": [
-        {
-            "name": "chocolate",
-            "index": [0, 1]
-        },
-        {
-            "name": "brick",
-            "index": [1, 0]
-        },
-        {
-            "name": "ground",
-            "index": [0, 0]
-        },
-        {
-            "name": "powerup-1",
-            "index": [24, 0]
-        },
-        {
-            "name": "powerup-2",
-            "index": [25, 0]
-        },
-        {
-            "name": "powerup-3",
-            "index": [26, 0]
-        },
-        {
-            "name": "sky",
-            "index": [3, 23]
-        }
-    ],
-    "animations": [
-        {
-            "name": "powerup",
-            "frameLength": 0.1,
-            "frames": [
-                "powerup-1",
-                "powerup-1",
-                "powerup-1",
-                "powerup-2",
-                "powerup-3",
-                "powerup-2"
-            ]
-        }
+  "imageURL": "img/tiles.png",
+  "tileW": 16,
+  "tileH": 16,
+  "tiles": [{
+      "name": "brick",
+      "index": [1, 0]
+    },
+    {
+      "name": "chocolate",
+      "index": [0, 1]
+    },
+    {
+      "name": "cloud-1-1",
+      "index": [0, 20]
+    },
+    {
+      "name": "cloud-1-2",
+      "index": [1, 20]
+    },
+    {
+      "name": "cloud-1-3",
+      "index": [2, 20]
+    },
+    {
+      "name": "cloud-2-1",
+      "index": [0, 21]
+    },
+    {
+      "name": "cloud-2-2",
+      "index": [1, 21]
+    },
+    {
+      "name": "cloud-2-3",
+      "index": [2, 21]
+    },
+    {
+      "name": "ground",
+      "index": [0, 0]
+    },
+    {
+      "name": "pipe-vert-cap-left",
+      "index": [0, 8]
+    },
+    {
+      "name": "pipe-vert-cap-right",
+      "index": [1, 8]
+    },
+    {
+      "name": "pipe-vert-section-left",
+      "index": [0, 9]
+    },
+    {
+      "name": "pipe-vert-section-right",
+      "index": [1, 9]
+    },
+    {
+      "name": "powerup-1",
+      "index": [24, 0]
+    },
+    {
+      "name": "powerup-2",
+      "index": [25, 0]
+    },
+    {
+      "name": "powerup-3",
+      "index": [26, 0]
+    },
+    {
+      "name": "sky",
+      "index": [3, 23]
+    }
+  ],
+  "animations": [{
+    "name": "powerup",
+    "frameLength": 0.1,
+    "frames": [
+      "powerup-1",
+      "powerup-1",
+      "powerup-1",
+      "powerup-2",
+      "powerup-3",
+      "powerup-2"
     ]
+  }]
 }
