+++ public/sprites/underworld.json
@@ -1,49 +1,86 @@
 {
-    "imageURL": "img/tiles.png",
-    "tileW": 16,
-    "tileH": 16,
-    "tiles": [
-        {
-            "name": "chocolate",
-            "index": [0, 3]
-        },
-        {
-            "name": "brick",
-            "index": [1, 2]
-        },
-        {
-            "name": "ground",
-            "index": [0, 2]
-        },
-        {
-            "name": "powerup-1",
-            "index": [24, 2]
-        },
-        {
-            "name": "powerup-2",
-            "index": [25, 2]
-        },
-        {
-            "name": "powerup-3",
-            "index": [26, 2]
-        },
-        {
-            "name": "sky",
-            "index": [13, 3]
-        }
-    ],
-    "animations": [
-        {
-            "name": "powerup",
-            "frameLength": 0.1,
-            "frames": [
-                "powerup-1",
-                "powerup-1",
-                "powerup-1",
-                "powerup-2",
-                "powerup-3",
-                "powerup-2"
-            ]
-        }
+  "imageURL": "img/tiles.png",
+  "tileW": 16,
+  "tileH": 16,
+  "tiles": [{
+      "name": "brick",
+      "index": [1, 2]
+    },
+    {
+      "name": "chocolate",
+      "index": [0, 3]
+    },
+    {
+      "name": "cloud-1-1",
+      "index": [0, 26]
+    },
+    {
+      "name": "cloud-1-2",
+      "index": [1, 26]
+    },
+    {
+      "name": "cloud-1-3",
+      "index": [2, 26]
+    },
+    {
+      "name": "cloud-2-1",
+      "index": [0, 27]
+    },
+    {
+      "name": "cloud-2-2",
+      "index": [1, 27]
+    },
+    {
+      "name": "cloud-2-3",
+      "index": [2, 27]
+    },
+    {
+      "name": "ground",
+      "index": [0, 2]
+    },
+    {
+      "name": "pipe-vert-cap-left",
+      "index": [0, 18]
+    },
+    {
+      "name": "pipe-vert-cap-right",
+      "index": [1, 18]
+    },
+    {
+      "name": "pipe-vert-section-left",
+      "index": [0, 19]
+    },
+    {
+      "name": "pipe-vert-section-right",
+      "index": [1, 19]
+    },
+    {
+      "name": "powerup-1",
+      "index": [24, 2]
+    },
+    {
+      "name": "powerup-2",
+      "index": [25, 2]
+    },
+    {
+      "name": "powerup-3",
+      "index": [26, 2]
+    },
+    {
+      "name": "sky",
+      "index": [13, 3]
+    }
+  ],
+  "animations": [{
+    "name": "powerup",
+    "frameLength": 0.1,
+    "frames": [
+      "powerup-1",
+      "powerup-1",
+      "powerup-1",
+      "powerup-2",
+      "powerup-3",
+      "powerup-2"
     ]
+  }]
 }
